Move the SHOW TABLE EXTENDED test out from the common trait
